pyproject.toml
[tool.poetry]
name = "jobsearch"
version = "0.1.0"
description = "CamelAI Agent workforce to search Python developer jobs and prepare interview plan"
authors = ["mvishiu11 <jakubmuszynski2103@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
rich = "^14.0.0"
pydantic-settings = "^2.9.1"
linkup-sdk = "^0.2.5"
camel-ai = {extras = ["all"], version = "^0.2.59"}
humanlayer = "^0.7.7"

[tool.poetry.scripts]
app = "jobsearch.run_agents:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
